requirements.txt
python-telegram-bot==20.7
Flask==3.1.2
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.1
psycopg2-binary==2.9.9
